Update code based on clang-tidy-14

+++ tesseract_geometry/include/tesseract_geometry/impl/mesh.h
@@ -89,7 +89,7 @@ public:
                   std::move(mesh_textures),
                   GeometryType::MESH)
   {
-    if ((getFaceCount() * 4) != getFaces()->size())
+    if ((static_cast<long>(getFaceCount()) * 4) != getFaces()->size())
       std::throw_with_nested(std::runtime_error("Mesh is not triangular"));  // LCOV_EXCL_LINE
   }
 
@@ -128,7 +128,7 @@ public:
                   std::move(mesh_textures),
                   GeometryType::MESH)
   {
-    if ((getFaceCount() * 4) != getFaces()->size())
+    if ((static_cast<long>(getFaceCount()) * 4) != getFaces()->size())
       std::throw_with_nested(std::runtime_error("Mesh is not triangular"));  // LCOV_EXCL_LINE
   }
